rust/aho-corasick: use same 'Set' structure as C
The C implementation rolls its own set data structure, but the
Rust program reuses HashSet. The C data structure is a simple
bespoke vector (i.e., "dynamic array"), where as a HashSet is
a pretty heavy duty data structure with a lot more relative
overhead.

This commit effectively copies the C 'Set' structure to the
Rust program. This alone brings the Rust program almost in line
with the performance of the C program.

This was caught immediately upon running the program under a
profiler.