Now does run_count+1 runs and throws away run 0
Due to noticing that sometimes the first run would have significantly
higher power usage than the rest, determined that these are cases
where the data files are not in the disk's cache and thus more power
is used in reading them. Discarding the first run should reduce the
effect of this, if not eliminate it.